requirements.txt
Django==1.7.7
Unidecode==0.04.17
argparse==1.2.1
django-email-extras==0.3.1
django-extensions==1.5.2
django-forms-builder==0.12.2
future==0.9.0
ipython==3.1.0
psycopg2==2.6
python-gnupg==0.3.7
six==1.9.0
sphinx-me==0.3
wsgiref==0.1.2
